refactor: move dialog components outside of install form
