+++ packages/frontend/src/modules/app/components/dialogs/install-dialog/install-dialog.tsx
@@ -1,14 +1,16 @@
 import { installAppMutation } from '@/api-client/@tanstack/react-query.gen';
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/Dialog';
+import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/Dialog';
+import { ScrollArea } from '@/components/ui/ScrollArea';
 import { useAppStatus } from '@/modules/app/helpers/use-app-status';
 import type { AppInfo } from '@/types/app.types';
 import type { TranslatableError } from '@/types/error.types';
 import { useMutation } from '@tanstack/react-query';
 import type React from 'react';
+import { useId } from 'react';
 import toast from 'react-hot-toast';
 import { useTranslation } from 'react-i18next';
+import { InstallFormButtons } from '../../install-form-buttons/install-form-buttons';
 import { InstallForm } from '../../install-form/install-form';
-import { ScrollArea } from '@/components/ui/ScrollArea';
 
 interface IProps {
   info: AppInfo;
@@ -19,6 +21,7 @@ interface IProps {
 export const InstallDialog: React.FC<IProps> = ({ info, isOpen, onClose }) => {
   const { t } = useTranslation();
   const { setOptimisticStatus } = useAppStatus();
+  const formId = useId();
 
   const installMutation = useMutation({
     ...installAppMutation(),
@@ -38,12 +41,18 @@ export const InstallDialog: React.FC<IProps> = ({ info, isOpen, onClose }) => {
           <DialogTitle>{t('APP_INSTALL_FORM_TITLE', { name: info.name })}</DialogTitle>
         </DialogHeader>
         <ScrollArea maxHeight={500}>
-          <InstallForm
-            onSubmit={(data) => installMutation.mutate({ path: { id: info.id }, body: data })}
-            formFields={info.form_fields}
-            info={info}
-          />
+          <DialogDescription>
+            <InstallForm
+              onSubmit={(data) => installMutation.mutate({ path: { id: info.id }, body: data })}
+              formFields={info.form_fields}
+              info={info}
+              formId={formId}
+            />
+          </DialogDescription>
         </ScrollArea>
+        <DialogFooter>
+          <InstallFormButtons loading={installMutation.isPending} formId={formId} />
+        </DialogFooter>
       </DialogContent>
     </Dialog>
   );
